Audit: Consolidate and improve propagation module tests (#1766)
Consolidate 6 sets of duplicate/near-duplicate test files (removing 11 files),
fix an assertion-less test in Trace.ts, and rename vague test names to
descriptive behavioural names across 5 files. No production code changes.

Addresses #1766
