+++ test/propagate/Bias.ts
@@ -1,7 +1,17 @@
+/**
+ * Consolidated tests for Bias — verifying accumulateBias direction and
+ * magnitude, limitBias boundary conditions and clamping, and batch
+ * accumulation functions.
+ *
+ * Merged from Bias.ts and BiasCalculation.ts as part of Issue #1766
+ * (propagation module test audit).
+ */
 import {
   assert,
   assertAlmostEquals,
   assertEquals,
+  assertGreater,
+  assertLess,
   assertThrows,
 } from "@std/assert";
 import { NeuronState } from "../../src/architecture/CreatureState.ts";
@@ -32,14 +42,39 @@ function makeConfig(
   });
 }
 
-function makeNeuronState(): NeuronState {
-  return new NeuronState();
-}
+// ---------------------------------------------------------------------------
+// accumulateBias - direction and magnitude
+// ---------------------------------------------------------------------------
 
-// --- accumulateBias ---
+Deno.test("accumulateBias - positive delta increases totalBias", () => {
+  const ns = new NeuronState();
+  const config = createBackPropagationConfig({
+    generations: 0,
+    learningRate: 1,
+  });
+
+  // targetPreActivation=5, preActivation=2 => delta=3, targetBias=0+3=3
+  accumulateBias(ns, 5, 2, 0, config);
+
+  assertEquals(ns.count, 1);
+  assertGreater(ns.totalBias, 0, "Positive delta should increase totalBias");
+});
+
+Deno.test("accumulateBias - negative delta decreases totalBias", () => {
+  const ns = new NeuronState();
+  const config = createBackPropagationConfig({
+    generations: 0,
+    learningRate: 1,
+  });
+
+  accumulateBias(ns, 1, 4, 0, config);
+
+  assertEquals(ns.count, 1);
+  assertLess(ns.totalBias, 0, "Negative delta should decrease totalBias");
+});
 
 Deno.test("accumulateBias - accumulates bias delta correctly", () => {
-  const ns = makeNeuronState();
+  const ns = new NeuronState();
   const currentBias = 0.5;
   const preActivation = 1.0;
   const targetPreActivation = 1.5;
@@ -59,7 +94,7 @@ Deno.test("accumulateBias - accumulates bias delta correctly", () => {
 });
 
 Deno.test("accumulateBias - accumulates multiple samples", () => {
-  const ns = makeNeuronState();
+  const ns = new NeuronState();
   const config = makeConfig();
 
   accumulateBias(ns, 2.0, 1.0, 0.5, config); // delta=1.0, target=1.5
@@ -70,33 +105,58 @@ Deno.test("accumulateBias - accumulates multiple samples", () => {
   assertAlmostEquals(ns.totalAdjustedBias, 3.0, 1e-10);
 });
 
+Deno.test("accumulateBias - zero delta still accumulates with currentBias", () => {
+  const ns = new NeuronState();
+  accumulateBias(ns, 1.0, 1.0, 0.5, makeConfig());
+  assertEquals(ns.count, 1);
+  assertAlmostEquals(ns.totalBias, 0.5, 1e-10);
+});
+
+// ---------------------------------------------------------------------------
+// accumulateBias - non-finite handling
+// ---------------------------------------------------------------------------
+
 Deno.test("accumulateBias - skips NaN target", () => {
-  const ns = makeNeuronState();
+  const ns = new NeuronState();
   accumulateBias(ns, NaN, 1.0, 0.5, makeConfig());
   assertEquals(ns.count, 0);
 });
 
+Deno.test("accumulateBias - skips Infinity target", () => {
+  const ns = new NeuronState();
+  const config = createBackPropagationConfig({
+    generations: 0,
+    learningRate: 1,
+  });
+  accumulateBias(ns, Infinity, 1, 0, config);
+  assertEquals(ns.count, 0, "Should skip Infinity targetPreActivation");
+});
+
 Deno.test("accumulateBias - skips Infinity preActivation", () => {
-  const ns = makeNeuronState();
+  const ns = new NeuronState();
   accumulateBias(ns, 1.0, Infinity, 0.5, makeConfig());
   assertEquals(ns.count, 0);
 });
 
+Deno.test("accumulateBias - skips -Infinity preActivation", () => {
+  const ns = new NeuronState();
+  const config = createBackPropagationConfig({
+    generations: 0,
+    learningRate: 1,
+  });
+  accumulateBias(ns, 1, -Infinity, 0, config);
+  assertEquals(ns.count, 0, "Should skip -Infinity preActivation");
+});
+
 Deno.test("accumulateBias - skips NaN currentBias", () => {
-  const ns = makeNeuronState();
+  const ns = new NeuronState();
   accumulateBias(ns, 1.0, 1.0, NaN, makeConfig());
   assertEquals(ns.count, 0);
 });
 
-Deno.test("accumulateBias - zero delta still accumulates", () => {
-  const ns = makeNeuronState();
-  accumulateBias(ns, 1.0, 1.0, 0.5, makeConfig());
-  assertEquals(ns.count, 1);
-  // delta = 0, targetBias = 0.5
-  assertAlmostEquals(ns.totalBias, 0.5, 1e-10);
-});
-
-// --- limitBias ---
+// ---------------------------------------------------------------------------
+// limitBias - learning rate and clamping
+// ---------------------------------------------------------------------------
 
 Deno.test("limitBias - applies learning rate to difference", () => {
   const config = makeConfig({ learningRate: 0.1 });
@@ -104,6 +164,21 @@ Deno.test("limitBias - applies learning rate to difference", () => {
   const result = limitBias(1.5, 0.5, config);
   assertAlmostEquals(result, 0.6, 1e-10);
 });
+
+Deno.test("limitBias - learning rate scales adjustment", () => {
+  const config = createBackPropagationConfig({
+    learningRate: 0.5,
+    maximumBiasAdjustmentScale: 100,
+    limitBiasScale: 100000,
+  });
+  // target=3, current=1, diff=2, scaled=0.5*2=1 => result=2
+  const result = limitBias(3, 1, config);
+  assertAlmostEquals(result, 2, 1e-9);
+});
+
+// ---------------------------------------------------------------------------
+// limitBias - tiny values and thresholds
+// ---------------------------------------------------------------------------
 
 Deno.test("limitBias - returns 0 for tiny target bias", () => {
   const config = makeConfig({ plankConstant: 1e-7 });
@@ -117,23 +192,20 @@ Deno.test("limitBias - returns currentBias when difference is tiny", () => {
   assertEquals(result, 0.5);
 });
 
-Deno.test("limitBias - throws on non-finite targetBias", () => {
-  const config = makeConfig();
-  assertThrows(() => limitBias(NaN, 0.5, config));
-  assertThrows(() => limitBias(Infinity, 0.5, config));
-});
+// ---------------------------------------------------------------------------
+// limitBias - adjustment scale clamping
+// ---------------------------------------------------------------------------
 
-Deno.test("limitBias - clamps large adjustment", () => {
+Deno.test("limitBias - clamps large positive adjustment", () => {
   const config = makeConfig({
     learningRate: 1,
     maximumBiasAdjustmentScale: 2,
   });
-  // targetBias=100, currentBias=0, diff=100, but clamped to ±2
   const result = limitBias(100, 0, config);
   assertAlmostEquals(result, 2, 1e-10);
 });
 
-Deno.test("limitBias - clamps negative large adjustment", () => {
+Deno.test("limitBias - clamps large negative adjustment", () => {
   const config = makeConfig({
     learningRate: 1,
     maximumBiasAdjustmentScale: 2,
@@ -142,27 +214,57 @@ Deno.test("limitBias - clamps negative large adjustment", () => {
   assertAlmostEquals(result, -2, 1e-10);
 });
 
-Deno.test("limitBias - enforces limitBiasScale", () => {
+Deno.test("limitBias - large gradient clamped to maximumBiasAdjustmentScale", () => {
+  const config = createBackPropagationConfig({
+    learningRate: 0.5,
+    maximumBiasAdjustmentScale: 1,
+    limitBiasScale: 10000,
+  });
+  const result = limitBias(1000, 0, config);
+  assertAlmostEquals(result, 1, 1e-9);
+});
+
+// ---------------------------------------------------------------------------
+// limitBias - limitBiasScale enforcement
+// ---------------------------------------------------------------------------
+
+Deno.test("limitBias - enforces limitBiasScale ceiling", () => {
   const config = makeConfig({
     learningRate: 1,
     maximumBiasAdjustmentScale: 100_000,
     limitBiasScale: 5,
   });
-  // targetBias=50, currentBias=0 => difference=50, limitedBias=50
-  // But |50| >= limitBiasScale(5), so clamp to 5
   const result = limitBias(50, 0, config);
   assert(Math.abs(result) <= 5, `Expected |result| <= 5, got ${result}`);
 });
 
-// --- accumulateBiasBatch4Way ---
+Deno.test("limitBias - enforces limitBiasScale negative floor", () => {
+  const config = createBackPropagationConfig({
+    learningRate: 1,
+    maximumBiasAdjustmentScale: 100,
+    limitBiasScale: 5,
+  });
+  const result = limitBias(-10, -4, config);
+  assertLess(result, -4, "Should move toward negative target");
+});
+
+// ---------------------------------------------------------------------------
+// limitBias - non-finite rejection
+// ---------------------------------------------------------------------------
+
+Deno.test("limitBias - throws on non-finite targetBias", () => {
+  const config = makeConfig();
+  assertThrows(() => limitBias(NaN, 0.5, config));
+  assertThrows(() => limitBias(Infinity, 0.5, config));
+  assertThrows(() => limitBias(-Infinity, 0, config), Error, "finite");
+});
+
+// ---------------------------------------------------------------------------
+// accumulateBiasBatch4Way
+// ---------------------------------------------------------------------------
 
 Deno.test("accumulateBiasBatch4Way - accumulates 4 neurons", () => {
-  const nsArray = [
-    makeNeuronState(),
-    makeNeuronState(),
-    makeNeuronState(),
-    makeNeuronState(),
-  ];
+  const nsArray = Array.from({ length: 4 }, () => new NeuronState());
   const targets = [2.0, 3.0, 4.0, 5.0];
   const preActs = [1.0, 1.0, 1.0, 1.0];
   const biases = [0.0, 0.0, 0.0, 0.0];
@@ -171,7 +273,6 @@ Deno.test("accumulateBiasBatch4Way - accumulates 4 neurons", () => {
 
   for (let i = 0; i < 4; i++) {
     assertEquals(nsArray[i].count, 1);
-    // delta = targets[i] - preActs[i], targetBias = biases[i] + delta
     assertAlmostEquals(
       nsArray[i].totalBias,
       targets[i] - preActs[i],
@@ -181,28 +282,25 @@ Deno.test("accumulateBiasBatch4Way - accumulates 4 neurons", () => {
 });
 
 Deno.test("accumulateBiasBatch4Way - skips non-finite values per element", () => {
-  const nsArray = [
-    makeNeuronState(),
-    makeNeuronState(),
-    makeNeuronState(),
-    makeNeuronState(),
-  ];
+  const nsArray = Array.from({ length: 4 }, () => new NeuronState());
   const targets = [2.0, NaN, 4.0, 5.0];
   const preActs = [1.0, 1.0, Infinity, 1.0];
   const biases = [0.0, 0.0, 0.0, NaN];
 
   accumulateBiasBatch4Way(nsArray, targets, preActs, biases, makeConfig());
 
-  assertEquals(nsArray[0].count, 1); // valid
+  assertEquals(nsArray[0].count, 1);
   assertEquals(nsArray[1].count, 0); // NaN target
   assertEquals(nsArray[2].count, 0); // Infinity preActivation
   assertEquals(nsArray[3].count, 0); // NaN bias
 });
 
-// --- accumulateBiasBatch8Way ---
+// ---------------------------------------------------------------------------
+// accumulateBiasBatch8Way
+// ---------------------------------------------------------------------------
 
 Deno.test("accumulateBiasBatch8Way - accumulates 8 neurons", () => {
-  const nsArray = Array.from({ length: 8 }, () => makeNeuronState());
+  const nsArray = Array.from({ length: 8 }, () => new NeuronState());
   const targets = [1, 2, 3, 4, 5, 6, 7, 8];
   const preActs = [0, 0, 0, 0, 0, 0, 0, 0];
   const biases = [0, 0, 0, 0, 0, 0, 0, 0];
@@ -216,7 +314,7 @@ Deno.test("accumulateBiasBatch8Way - accumulates 8 neurons", () => {
 });
 
 Deno.test("accumulateBiasBatch8Way - skips non-finite values selectively", () => {
-  const nsArray = Array.from({ length: 8 }, () => makeNeuronState());
+  const nsArray = Array.from({ length: 8 }, () => new NeuronState());
   const targets = [1, NaN, 3, 4, 5, 6, 7, 8];
   const preActs = [0, 0, 0, 0, 0, 0, 0, 0];
   const biases = [0, 0, 0, 0, 0, 0, 0, 0];
@@ -224,6 +322,6 @@ Deno.test("accumulateBiasBatch8Way - skips non-finite values selectively", () =>
   accumulateBiasBatch8Way(nsArray, targets, preActs, biases, makeConfig());
 
   assertEquals(nsArray[0].count, 1);
-  assertEquals(nsArray[1].count, 0); // skipped due to NaN
+  assertEquals(nsArray[1].count, 0);
   assertEquals(nsArray[2].count, 1);
 });
