+++ test/propagate/RecordElasticity.ts
@@ -1,6 +1,17 @@
-import { assert, assertAlmostEquals, assertEquals } from "@std/assert";
+/**
+ * Consolidated tests for RecordElasticity — verifying elastic error
+ * distribution, feasibility factors, record value computation,
+ * link building, and constrained redistribution.
+ *
+ * Merged from RecordElasticity.ts, RecordElasticityBehaviour.ts,
+ * RecordElasticityTest.ts as part of Issue #1766 (propagation module
+ * test audit).
+ */
+import { assertAlmostEquals, assertEquals, assertGreater } from "@std/assert";
 import { Creature } from "../../src/Creature.ts";
 import type { CreatureExport } from "../../src/architecture/CreatureInterfaces.ts";
+import type { Neuron } from "../../src/architecture/Neuron.ts";
+import type { Synapse } from "../../src/architecture/Synapse.ts";
 import {
   buildRecordElasticLinks,
   constrainAndRedistributeRecordShares,
@@ -12,57 +23,173 @@ import {
 import { SparseConfig } from "../../src/propagate/sparse/SparseConfig.ts";
 import { createBackPropagationConfig } from "../../src/propagate/BackPropagation.ts";
 
-// --- distributeRecordError ---
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
 
-Deno.test("distributeRecordError - returns zeroes for empty links", () => {
+function makeLink(opts: {
+  weight: number;
+  fromActivation: number;
+  safeZoneFactor?: number;
+  feasibilityFactor?: number;
+  fromNeuron?: Neuron;
+}): RecordElasticLink {
+  const weight = opts.weight;
+  const fromActivation = opts.fromActivation;
+  return {
+    synapse: { from: 0, to: 1, weight } as Synapse,
+    fromNeuron: opts.fromNeuron ?? { type: "input" } as unknown as Neuron,
+    fromActivation,
+    fromValue: weight * fromActivation,
+    safeZoneFactor: opts.safeZoneFactor ?? 1,
+    feasibilityFactor: opts.feasibilityFactor ?? 1,
+  };
+}
+
+function makeCreatureWithSquash(squashName: string): {
+  creature: Creature;
+  hiddenNeuron: Neuron;
+} {
+  const json: CreatureExport = {
+    neurons: [
+      { type: "hidden", uuid: "hidden-0", squash: squashName, bias: 0 },
+      { type: "output", uuid: "output-0", squash: "IDENTITY", bias: 0 },
+    ],
+    synapses: [
+      { fromUUID: "input-0", toUUID: "hidden-0", weight: 1 },
+      { fromUUID: "hidden-0", toUUID: "output-0", weight: 1 },
+    ],
+    input: 1,
+    output: 1,
+  };
+  const creature = Creature.fromJSON(json);
+  creature.validate();
+  return { creature, hiddenNeuron: creature.neurons[1] };
+}
+
+// ---------------------------------------------------------------------------
+// distributeRecordError - basic behaviour
+// ---------------------------------------------------------------------------
+
+Deno.test("distributeRecordError - empty links returns empty shares", () => {
   const result = distributeRecordError(1.0, []);
   assertEquals(result.shares.length, 0);
 });
 
-Deno.test("distributeRecordError - returns zeroes for non-finite error", () => {
-  const link = makeMockLink(1.0, 0.5, 1.0, 1.0);
+Deno.test("distributeRecordError - non-finite error returns zero shares", () => {
+  const link = makeLink({ weight: 1.0, fromActivation: 0.5 });
   const result = distributeRecordError(NaN, [link]);
   assertEquals(result.shares[0], 0);
 });
 
 Deno.test("distributeRecordError - single link gets all error", () => {
-  const link = makeMockLink(1.0, 0.5, 1.0, 1.0);
-  const result = distributeRecordError(2.0, [link]);
-  assertAlmostEquals(result.shares[0], 2.0, 1e-10);
+  const link = makeLink({ weight: 2.0, fromActivation: 0.5 });
+  const result = distributeRecordError(3.0, [link]);
+  assertAlmostEquals(result.shares[0], 3.0, 1e-9);
 });
 
-Deno.test("distributeRecordError - distributes by weight-squared", () => {
-  // Links use synapse.weight for distribution (weight-squared elasticity)
-  const link1 = makeMockLink(2.0, 0.5, 1.0, 1.0); // weight=2
-  const link2 = makeMockLink(1.0, 0.5, 1.0, 1.0); // weight=1
+// ---------------------------------------------------------------------------
+// distributeRecordError - weight-squared proportional distribution
+// ---------------------------------------------------------------------------
 
-  const result = distributeRecordError(5.0, [link1, link2]);
-  const total = result.shares[0] + result.shares[1];
-  assertAlmostEquals(total, 5.0, 1e-10);
+Deno.test("distributeRecordError - distributes proportional to weight squared", () => {
+  const links = [
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+    makeLink({ weight: 2, fromActivation: 0.5 }),
+    makeLink({ weight: 3, fromActivation: 0.5 }),
+  ];
+  const result = distributeRecordError(14, links);
 
-  // weight² ratio: 4:1, so link1 gets 4/5, link2 gets 1/5
-  assert(
-    result.shares[0] > result.shares[1],
-    "Larger weight should get larger share",
-  );
+  // weight² scores: 1, 4, 9 => total 14
+  assertAlmostEquals(result.shares[0], 1, 1e-9);
+  assertAlmostEquals(result.shares[1], 4, 1e-9);
+  assertAlmostEquals(result.shares[2], 9, 1e-9);
 });
 
-Deno.test("distributeRecordError - safeZoneFactor blocks allocation", () => {
-  const open = makeMockLink(1.0, 0.5, 1.0, 1.0); // safeZone=1
-  const blocked = makeMockLink(1.0, 0.5, 0.0, 1.0); // safeZone=0
-
-  const result = distributeRecordError(3.0, [open, blocked]);
-  // Blocked link should get less (or zero) share
-  assert(
-    Math.abs(result.shares[0]) >= Math.abs(result.shares[1]),
-    "Blocked link should not dominate",
-  );
+Deno.test("distributeRecordError - shares sum to original error", () => {
+  const links = [
+    makeLink({ weight: 1.5, fromActivation: 0.3 }),
+    makeLink({ weight: 2.5, fromActivation: 0.7 }),
+    makeLink({ weight: 0.5, fromActivation: 0.1 }),
+  ];
+  const result = distributeRecordError(7.5, links);
+  const sum = result.shares.reduce((a, b) => a + b, 0);
+  assertAlmostEquals(sum, 7.5, 1e-9, "Shares should sum to original error");
 });
+
+Deno.test("distributeRecordError - negative error distributes correctly", () => {
+  const links = [
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+    makeLink({ weight: 3, fromActivation: 0.5 }),
+  ];
+  const result = distributeRecordError(-10, links);
+
+  // Same weight² proportions but negative: 1:9
+  assertAlmostEquals(result.shares[0], -1, 1e-9);
+  assertAlmostEquals(result.shares[1], -9, 1e-9);
+});
+
+// ---------------------------------------------------------------------------
+// distributeRecordError - gating via safeZoneFactor and feasibilityFactor
+// ---------------------------------------------------------------------------
+
+Deno.test("distributeRecordError - safeZoneFactor gates contribution", () => {
+  const links = [
+    makeLink({ weight: 2, fromActivation: 0.5, safeZoneFactor: 1 }),
+    makeLink({ weight: 2, fromActivation: 0.5, safeZoneFactor: 0 }),
+  ];
+  const result = distributeRecordError(10, links);
+
+  assertAlmostEquals(result.shares[0], 10, 1e-6);
+  assertAlmostEquals(result.shares[1], 0, 1e-6);
+});
+
+Deno.test("distributeRecordError - feasibilityFactor gates contribution", () => {
+  const links = [
+    makeLink({ weight: 2, fromActivation: 0.5, feasibilityFactor: 1 }),
+    makeLink({
+      weight: 2,
+      fromActivation: 0.5,
+      safeZoneFactor: 1,
+      feasibilityFactor: 0,
+    }),
+  ];
+  const result = distributeRecordError(10, links);
+
+  assertAlmostEquals(result.shares[0], 10, 1e-6);
+  assertAlmostEquals(result.shares[1], 0, 1e-6);
+});
+
+Deno.test("distributeRecordError - both gates combine multiplicatively", () => {
+  const links = [
+    makeLink({
+      weight: 2,
+      fromActivation: 0.5,
+      safeZoneFactor: 0.5,
+      feasibilityFactor: 0.5,
+    }),
+    makeLink({
+      weight: 2,
+      fromActivation: 0.5,
+      safeZoneFactor: 1.0,
+      feasibilityFactor: 1.0,
+    }),
+  ];
+  const result = distributeRecordError(10, links);
+
+  // Gate scores: 0.5*0.5=0.25 vs 1.0*1.0=1.0
+  // Weighted scores: 4*0.25=1 vs 4*1.0=4 => shares: 2 vs 8
+  assertAlmostEquals(result.shares[0], 2, 1e-9);
+  assertAlmostEquals(result.shares[1], 8, 1e-9);
+});
+
+// ---------------------------------------------------------------------------
+// distributeRecordError - fallback behaviour
+// ---------------------------------------------------------------------------
 
 Deno.test("distributeRecordError - all blocked falls back to equal split", () => {
-  const a = makeMockLink(1.0, 0.0, 0.0, 0.0);
-  const b = makeMockLink(1.0, 0.0, 0.0, 0.0);
-
+  const a = makeLink({ weight: 1.0, fromActivation: 0.0, safeZoneFactor: 0 });
+  const b = makeLink({ weight: 1.0, fromActivation: 0.0, safeZoneFactor: 0 });
   const result = distributeRecordError(4.0, [a, b], {
     allowEqualFallback: true,
   });
@@ -71,17 +198,18 @@ Deno.test("distributeRecordError - all blocked falls back to equal split", () =>
 });
 
 Deno.test("distributeRecordError - no equal fallback returns zeroes", () => {
-  const a = makeMockLink(1.0, 0.0, 0.0, 0.0);
-  const b = makeMockLink(1.0, 0.0, 0.0, 0.0);
-
+  const a = makeLink({ weight: 1.0, fromActivation: 0.0, safeZoneFactor: 0 });
+  const b = makeLink({ weight: 1.0, fromActivation: 0.0, safeZoneFactor: 0 });
   const result = distributeRecordError(4.0, [a, b], {
     allowEqualFallback: false,
   });
-  assertEquals(result.shares[0], 0);
-  assertEquals(result.shares[1], 0);
+  assertAlmostEquals(result.shares[0], 0, 1e-12);
+  assertAlmostEquals(result.shares[1], 0, 1e-12);
 });
 
-// --- recordTargetFeasibilityFactor ---
+// ---------------------------------------------------------------------------
+// recordTargetFeasibilityFactor - non-finite handling
+// ---------------------------------------------------------------------------
 
 Deno.test("recordTargetFeasibilityFactor - returns 0 for non-finite activation", () => {
   const json: CreatureExport = {
@@ -95,85 +223,66 @@ Deno.test("recordTargetFeasibilityFactor - returns 0 for non-finite activation",
     ],
   };
   const creature = Creature.fromJSON(json);
-  const neuron = creature.neurons[1]; // output neuron
+  const neuron = creature.neurons[1];
 
   assertEquals(recordTargetFeasibilityFactor(neuron, NaN), 0);
   assertEquals(recordTargetFeasibilityFactor(neuron, Infinity), 0);
+  assertEquals(recordTargetFeasibilityFactor(neuron, -Infinity), 0);
 });
 
-Deno.test("recordTargetFeasibilityFactor - returns 1 for IDENTITY in range", () => {
-  const json: CreatureExport = {
-    input: 1,
-    output: 1,
-    neurons: [
-      { type: "output", uuid: "output-0", squash: "IDENTITY", bias: 0 },
-    ],
-    synapses: [
-      { fromUUID: "input-0", toUUID: "output-0", weight: 1 },
-    ],
-  };
-  const creature = Creature.fromJSON(json);
-  const neuron = creature.neurons[1];
+// ---------------------------------------------------------------------------
+// recordTargetFeasibilityFactor - per-squash behaviour
+// ---------------------------------------------------------------------------
 
-  assertEquals(recordTargetFeasibilityFactor(neuron, 0.5), 1);
+Deno.test("recordTargetFeasibilityFactor - IDENTITY always returns 1", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("IDENTITY");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, 100),
+    1,
+    1e-9,
+  );
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, -100),
+    1,
+    1e-9,
+  );
+  assertAlmostEquals(recordTargetFeasibilityFactor(hiddenNeuron, 0), 1, 1e-9);
 });
 
 Deno.test("recordTargetFeasibilityFactor - ReLU returns 0.1 for negative target", () => {
-  const json: CreatureExport = {
-    input: 1,
-    output: 1,
-    neurons: [
-      { type: "hidden", uuid: "h1", squash: "ReLU", bias: 0 },
-      { type: "output", uuid: "output-0", squash: "IDENTITY", bias: 0 },
-    ],
-    synapses: [
-      { fromUUID: "input-0", toUUID: "h1", weight: 1 },
-      { fromUUID: "h1", toUUID: "output-0", weight: 1 },
-    ],
-  };
-  const creature = Creature.fromJSON(json);
-  const reluNeuron = creature.neurons[1]; // hidden neuron with ReLU
-
-  // ReLU range is [0, ∞), negative target is out of range
-  assertEquals(recordTargetFeasibilityFactor(reluNeuron, -1.0), 0.1);
+  const { hiddenNeuron } = makeCreatureWithSquash("ReLU");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, -1.0),
+    0.1,
+    1e-9,
+  );
 });
 
 Deno.test("recordTargetFeasibilityFactor - ReLU returns 1 for positive target", () => {
-  const json: CreatureExport = {
-    input: 1,
-    output: 1,
-    neurons: [
-      { type: "hidden", uuid: "h1", squash: "ReLU", bias: 0 },
-      { type: "output", uuid: "output-0", squash: "IDENTITY", bias: 0 },
-    ],
-    synapses: [
-      { fromUUID: "input-0", toUUID: "h1", weight: 1 },
-      { fromUUID: "h1", toUUID: "output-0", weight: 1 },
-    ],
-  };
-  const creature = Creature.fromJSON(json);
-  const reluNeuron = creature.neurons[1];
-
-  assertEquals(recordTargetFeasibilityFactor(reluNeuron, 5.0), 1);
+  const { hiddenNeuron } = makeCreatureWithSquash("ReLU");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, 5.0),
+    1,
+    1e-9,
+  );
 });
 
 Deno.test("recordTargetFeasibilityFactor - Swish returns 0.1 for negative target", () => {
-  const json: CreatureExport = {
-    input: 1,
-    output: 1,
-    neurons: [
-      { type: "hidden", uuid: "h1", squash: "Swish", bias: 0 },
-      { type: "output", uuid: "output-0", squash: "IDENTITY", bias: 0 },
-    ],
-    synapses: [
-      { fromUUID: "input-0", toUUID: "h1", weight: 1 },
-      { fromUUID: "h1", toUUID: "output-0", weight: 1 },
-    ],
-  };
-  const creature = Creature.fromJSON(json);
-  const swishNeuron = creature.neurons[1];
+  const { hiddenNeuron } = makeCreatureWithSquash("Swish");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, -1.0),
+    0.1,
+    1e-9,
+  );
+});
 
-  assertEquals(recordTargetFeasibilityFactor(swishNeuron, -1.0), 0.1);
+Deno.test("recordTargetFeasibilityFactor - Swish returns 1 for positive target", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("Swish");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, 0.5),
+    1,
+    1e-9,
+  );
 });
 
 Deno.test("recordTargetFeasibilityFactor - TANH returns 1 for in-range target", () => {
@@ -191,11 +300,75 @@ Deno.test("recordTargetFeasibilityFactor - TANH returns 1 for in-range target", 
   };
   const creature = Creature.fromJSON(json);
   const tanhNeuron = creature.neurons[1];
-
   assertEquals(recordTargetFeasibilityFactor(tanhNeuron, 0.5), 1);
 });
 
-// --- getOrComputeRecordValue ---
+Deno.test("recordTargetFeasibilityFactor - ABSOLUTE returns 0.1 for negative target", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("ABSOLUTE");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, -1),
+    0.1,
+    1e-9,
+  );
+});
+
+Deno.test("recordTargetFeasibilityFactor - SQUARE out-of-range returns 0.1", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("SQUARE");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, -1),
+    0.1,
+    1e-9,
+  );
+});
+
+Deno.test("recordTargetFeasibilityFactor - SQUARE in-range returns 1", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("SQUARE");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, 0.5),
+    1,
+    1e-9,
+  );
+});
+
+Deno.test("recordTargetFeasibilityFactor - SELU negative returns 0.1", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("SELU");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, -0.5),
+    0.1,
+    1e-9,
+  );
+});
+
+Deno.test("recordTargetFeasibilityFactor - SELU positive returns 1", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("SELU");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, 0.5),
+    1,
+    1e-9,
+  );
+});
+
+Deno.test("recordTargetFeasibilityFactor - GELU negative returns 0.1", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("GELU");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, -0.5),
+    0.1,
+    1e-9,
+  );
+});
+
+Deno.test("recordTargetFeasibilityFactor - ELU negative returns 0.1", () => {
+  const { hiddenNeuron } = makeCreatureWithSquash("ELU");
+  assertAlmostEquals(
+    recordTargetFeasibilityFactor(hiddenNeuron, -0.5),
+    0.1,
+    1e-9,
+  );
+});
+
+// ---------------------------------------------------------------------------
+// getOrComputeRecordValue
+// ---------------------------------------------------------------------------
 
 Deno.test("getOrComputeRecordValue - returns existing value from discoverMap", () => {
   const json: CreatureExport = {
@@ -260,7 +433,9 @@ Deno.test("getOrComputeRecordValue - computes value for input neuron", () => {
   assertAlmostEquals(result, 0.7, 1e-5);
 });
 
-// --- buildRecordElasticLinks ---
+// ---------------------------------------------------------------------------
+// buildRecordElasticLinks
+// ---------------------------------------------------------------------------
 
 Deno.test("buildRecordElasticLinks - builds links from inward synapses", () => {
   const json: CreatureExport = {
@@ -285,7 +460,11 @@ Deno.test("buildRecordElasticLinks - builds links from inward synapses", () => {
   });
   const sparseConfig = new SparseConfig(json, config);
 
-  creature.activateAndTrace(new Float32Array([1.0, 0.5]), false, sparseConfig);
+  creature.activateAndTrace(
+    new Float32Array([1.0, 0.5]),
+    false,
+    sparseConfig,
+  );
 
   const outputNeuron = creature.neurons[creature.neurons.length - 1];
   const inward = creature.inwardConnections(outputNeuron.index);
@@ -298,14 +477,13 @@ Deno.test("buildRecordElasticLinks - builds links from inward synapses", () => {
     0.1,
   );
 
-  // Should have 2 links (from h1 and h2, not from inputs by default)
   assertEquals(links.length, 2);
   for (const link of links) {
-    assert(link.synapse !== undefined);
-    assert(link.fromNeuron !== undefined);
-    assert(Number.isFinite(link.fromActivation));
-    assert(Number.isFinite(link.safeZoneFactor));
-    assert(Number.isFinite(link.feasibilityFactor));
+    assertEquals(link.synapse !== undefined, true);
+    assertEquals(link.fromNeuron !== undefined, true);
+    assertEquals(Number.isFinite(link.fromActivation), true);
+    assertEquals(Number.isFinite(link.safeZoneFactor), true);
+    assertEquals(Number.isFinite(link.feasibilityFactor), true);
   }
 });
 
@@ -333,7 +511,7 @@ Deno.test("buildRecordElasticLinks - includeInputNodes option", () => {
   const inward = creature.inwardConnections(outputNeuron.index);
   const discoverMap = new Map();
 
-  // Without includeInputNodes (default false)
+  // Default: input nodes excluded
   const linksDefault = buildRecordElasticLinks(
     outputNeuron,
     inward,
@@ -353,16 +531,18 @@ Deno.test("buildRecordElasticLinks - includeInputNodes option", () => {
   assertEquals(linksIncluding.length, 1);
 });
 
-// --- constrainAndRedistributeRecordShares ---
+// ---------------------------------------------------------------------------
+// constrainAndRedistributeRecordShares
+// ---------------------------------------------------------------------------
 
-Deno.test("constrainAndRedistributeRecordShares - passes through unconstrained shares", () => {
-  const link = makeMockLink(1.0, 0.5, 1.0, 1.0);
-  const shares = constrainAndRedistributeRecordShares(
-    1.0,
-    [link],
-    [1.0],
-  );
-  assertAlmostEquals(shares[0], 1.0, 1e-10);
+Deno.test("constrainAndRedistributeRecordShares - unconstrained shares pass through", () => {
+  const links = [
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+  ];
+  const result = constrainAndRedistributeRecordShares(10, links, [4, 6]);
+  assertAlmostEquals(result[0], 4, 1e-9);
+  assertAlmostEquals(result[1], 6, 1e-9);
 });
 
 Deno.test("constrainAndRedistributeRecordShares - handles empty links", () => {
@@ -370,33 +550,131 @@ Deno.test("constrainAndRedistributeRecordShares - handles empty links", () => {
   assertEquals(shares.length, 0);
 });
 
-// --- Helper to create mock RecordElasticLink ---
+Deno.test("constrainAndRedistributeRecordShares - zero-weight link gets zero share", () => {
+  const links = [
+    makeLink({ weight: 0, fromActivation: 0.5 }),
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+  ];
+  const result = constrainAndRedistributeRecordShares(10, links, [5, 5]);
+  assertAlmostEquals(result[0], 0, 1e-9);
+  assertAlmostEquals(result[1], 10, 1e-6);
+});
 
-function makeMockLink(
-  weight: number,
-  fromActivation: number,
-  safeZoneFactor: number,
-  feasibilityFactor: number,
-): RecordElasticLink {
-  // Create a minimal mock that satisfies the RecordElasticLink type
-  return {
-    synapse: {
-      weight,
-      from: 0,
-      to: 1,
-    } as unknown as import("../../src/architecture/Synapse.ts").Synapse,
-    fromNeuron: {
-      type: "hidden",
-      index: 0,
-      uuid: "mock",
-      findSquash: () => ({
-        range: { low: -1, high: 1 },
-        getName: () => "IDENTITY",
-      }),
-    } as unknown as import("../../src/architecture/Neuron.ts").Neuron,
-    fromActivation,
-    fromValue: weight * fromActivation,
-    safeZoneFactor,
-    feasibilityFactor,
+Deno.test("constrainAndRedistributeRecordShares - blocked links redistribute to open links", () => {
+  const links = [
+    makeLink({
+      weight: 1,
+      fromActivation: 0.5,
+      safeZoneFactor: 0,
+      feasibilityFactor: 0,
+    }),
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+  ];
+  const result = constrainAndRedistributeRecordShares(10, links, [3, 3, 4]);
+
+  assertAlmostEquals(result[0], 0, 1e-9);
+  const openSum = result[1] + result[2];
+  assertAlmostEquals(openSum, 10, 1e-6, "Open links should absorb all error");
+});
+
+Deno.test("constrainAndRedistributeRecordShares - all blocked results in zero shares", () => {
+  const links = [
+    makeLink({
+      weight: 1,
+      fromActivation: 0.5,
+      safeZoneFactor: 0,
+      feasibilityFactor: 0,
+    }),
+    makeLink({
+      weight: 1,
+      fromActivation: 0.5,
+      safeZoneFactor: 0,
+      feasibilityFactor: 0,
+    }),
+  ];
+  const result = constrainAndRedistributeRecordShares(10, links, [5, 5]);
+  assertAlmostEquals(result[0], 0, 1e-9);
+  assertAlmostEquals(result[1], 0, 1e-9);
+});
+
+Deno.test("constrainAndRedistributeRecordShares - shares sum preserved", () => {
+  const links = [
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+    makeLink({ weight: 2, fromActivation: 0.5 }),
+    makeLink({ weight: 3, fromActivation: 0.5 }),
+  ];
+  const result = constrainAndRedistributeRecordShares(10, links, [2, 3, 5]);
+  const sum = result.reduce((a, b) => a + b, 0);
+  assertAlmostEquals(sum, 10, 1e-6, "Total shares should be preserved");
+});
+
+Deno.test("constrainAndRedistributeRecordShares - handles negative error", () => {
+  const links = [
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+    makeLink({ weight: 1, fromActivation: 0.5 }),
+  ];
+  const result = constrainAndRedistributeRecordShares(-10, links, [-3, -7]);
+  const sum = result.reduce((a, b) => a + b, 0);
+  assertAlmostEquals(sum, -10, 1e-6);
+});
+
+// ---------------------------------------------------------------------------
+// constrainAndRedistributeRecordShares - with real creature
+// ---------------------------------------------------------------------------
+
+Deno.test("constrainAndRedistributeRecordShares - redistributes away from range-limited squash", () => {
+  const json: CreatureExport = {
+    neurons: [
+      { type: "hidden", uuid: "hidden-abs", squash: "ABSOLUTE", bias: 0 },
+      { type: "hidden", uuid: "hidden-id", squash: "IDENTITY", bias: 0 },
+      { type: "output", uuid: "output-0", squash: "IDENTITY", bias: 0 },
+    ],
+    synapses: [
+      { fromUUID: "input-0", toUUID: "hidden-abs", weight: 1 },
+      { fromUUID: "input-0", toUUID: "hidden-id", weight: 1 },
+      { fromUUID: "hidden-abs", toUUID: "output-0", weight: 1 },
+      { fromUUID: "hidden-id", toUUID: "output-0", weight: 1 },
+    ],
+    input: 1,
+    output: 1,
   };
-}
+  const creature = Creature.fromJSON(json);
+  creature.validate();
+  creature.activate(new Float32Array([0.5]));
+
+  const absNeuron = creature.neurons[1];
+  const idNeuron = creature.neurons[2];
+
+  const links: RecordElasticLink[] = [
+    {
+      synapse: creature.synapses[2],
+      fromNeuron: absNeuron,
+      fromActivation: 0.5,
+      fromValue: 0.5,
+      safeZoneFactor: 1,
+      feasibilityFactor: recordTargetFeasibilityFactor(absNeuron, -1),
+    },
+    {
+      synapse: creature.synapses[3],
+      fromNeuron: idNeuron,
+      fromActivation: 0.5,
+      fromValue: 0.5,
+      safeZoneFactor: 1,
+      feasibilityFactor: recordTargetFeasibilityFactor(idNeuron, -1),
+    },
+  ];
+
+  // ABSOLUTE neuron should have lower feasibility for negative target
+  assertAlmostEquals(links[0].feasibilityFactor, 0.1, 1e-9);
+  assertAlmostEquals(links[1].feasibilityFactor, 1, 1e-9);
+
+  const result = distributeRecordError(-2, links);
+
+  // IDENTITY link should absorb more negative error
+  assertGreater(
+    Math.abs(result.shares[1]),
+    Math.abs(result.shares[0]),
+    "IDENTITY link should absorb more error than ABSOLUTE link",
+  );
+});
